requirements.txt
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1
PyYAML==6.0